requirements.txt
faiss-cpu
sentence-transformers
schedule
